flask/requirements.txt
argh==0.26.2
attrs==20.3.0
autopep8==1.5.4
click==7.1.2
Flask==1.1.0
icdiff==1.9.1
importlib-metadata==3.4.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
packaging==20.8
pluggy==0.13.1
pprintpp==0.4.0
py==1.10.0
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==6.2.2
pytest-icdiff==0.5
PyYAML==5.4.1
toml==0.10.2
typing-extensions==3.7.4.3
uWSGI==2.0.19.1
watchdog==1.0.2
Werkzeug==1.0.1
zipp==3.4.0
